share settings introduce tabs
