+++ projects/frontend/src/components/HeaderMenu.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react"
+import React, { useState, useEffect } from "react"
 import { PlusOutlined, ShareAltOutlined, ProfileOutlined } from "@ant-design/icons"
 import { Menu } from "antd"
 import styled from "styled-components"
@@ -165,7 +165,7 @@ export const HeaderNavMenu = () => {
 					onCancel={() => setShowCreateShare(false)}
 				/>
 			)}
-			{shareSettings && <ShareSettings share={shareSettings} onClose={() => setShareSettings(null)} />}
+			{libraryShare && <ShareSettings share={libraryShare} onClose={() => setShareSettings(null)} />}
 			{showChangePassword && <ChangePasswordModal onClose={() => setShowChangePassword(false)} />}
 		</>
 	)
