+++ projects/frontend/src/components/modals/share-settings/ShareSettings.tsx
@@ -1,32 +1,15 @@
-import React, { useState, useEffect, useCallback, useMemo } from "react"
+import React, { useCallback, useMemo, useState } from "react"
 import { QuestionCircleOutlined } from "@ant-design/icons"
-import { Modal, Input, Table, Button, Alert, Popconfirm, Typography, message, Form } from "antd"
-import { useDebounce } from "use-debounce/lib"
-import {
-	useShareUsers,
-	useUpdateShareMemberPermissions,
-	useRevokeInvitation,
-	useInviteToShare,
-	useRenameShare,
-	useLeaveShare,
-	useDeleteShare,
-} from "@musicshare/react-graphql-client"
-import Column from "antd/lib/table/Column"
-import { Permissions, UserStatus, ShareMember, Permission, Share } from "@musicshare/shared-types"
+import { Modal, Button, Popconfirm, message, Tabs } from "antd"
+import { useLeaveShare, useDeleteShare } from "@musicshare/react-graphql-client"
+import { Permissions, Share } from "@musicshare/shared-types"
 import { useHistory } from "react-router-dom"
-import styled from "styled-components"
-import { EditableTagGroup } from "../../form/EditableTagGroup"
 import { useLibraryID } from "../../../hooks/data/useLibraryID"
+import { ShareSettingsGeneral } from "./ShareSettingsGeneral"
 
-const { Text } = Typography
+const { TabPane } = Tabs
 
-const FormItemVertical = styled(Form.Item)`
-	flex-direction: column;
-
-	& .ant-col {
-		text-align: left;
-	}
-`
+type ShareSettingsTab = "general" | "metadata"
 
 interface IShareSettingsProps {
 	share: Share
@@ -34,6 +17,7 @@ interface IShareSettingsProps {
 }
 
 export const ShareSettings: React.FC<IShareSettingsProps> = ({ share, onClose }) => {
+	const [tab, setTab] = useState<ShareSettingsTab>("general")
 	const history = useHistory()
 	const [deleteShare] = useDeleteShare({
 		onSuccess: () => {
@@ -52,8 +36,6 @@ export const ShareSettings: React.FC<IShareSettingsProps> = ({ share, onClose })
 	const userLibraryID = useLibraryID()
 	const isLibrary = share.id === userLibraryID
 	const isOwner = useMemo(() => share.userPermissions.includes(Permissions.SHARE_OWNER), [share.userPermissions])
-	const canChangeName = isLibrary || isOwner
-	const canInvite = !isLibrary && isOwner
 
 	const onLeaveDeleteClick = useCallback(() => {
 		if (isOwner) {
@@ -91,181 +73,16 @@ export const ShareSettings: React.FC<IShareSettingsProps> = ({ share, onClose })
 				onClick: (e) => e.preventDefault(),
 			}}
 		>
-			<Form>
-				{canChangeName && <ChangeSongName share={share} />}
-				{canInvite && <ShareUsers shareID={share.id} />}
-				{!isOwner && <div>{"You've missing the required permission to edit share settings"}</div>}
-			</Form>
+			<Tabs defaultActiveKey={tab} activeKey={tab} onChange={(tabKey) => setTab(tabKey as ShareSettingsTab)}>
+				<TabPane tab="Generak" key="general">
+					<ShareSettingsGeneral share={share} isLibrary={isLibrary} isOwner={isOwner} />
+				</TabPane>
+				{isLibrary && (
+					<TabPane tab="Meta Data" key="metadata">
+						Content of Tab Pane 2
+					</TabPane>
+				)}
+			</Tabs>
 		</Modal>
-	)
-}
-
-const ChangeSongName: React.FC<{ share: Share }> = ({ share: { name, id } }) => {
-	const [shareName, setShareName] = useState(name)
-	const [inputBlured, setInputBlured] = useState(false)
-	const [debouncedShareName] = useDebounce(shareName, 1000)
-	const [renameShare] = useRenameShare()
-
-	useEffect(() => {
-		if (inputBlured) {
-			renameShare({
-				shareID: id,
-				name: debouncedShareName,
-			})
-		}
-	}, [debouncedShareName, id, renameShare, inputBlured])
-
-	return (
-		<FormItemVertical label="Name" validateStatus={shareName.trim().length <= 2 ? "error" : "success"}>
-			<Input
-				value={shareName}
-				type="text"
-				onChange={(e) => setShareName(e.target.value)}
-				placeholder="Share name"
-				onBlur={() => setInputBlured(true)}
-			/>
-		</FormItemVertical>
-	)
-}
-
-const ShareUsers: React.FC<{ shareID: string }> = ({ shareID }) => {
-	const { data: users, isLoading, error } = useShareUsers(shareID)
-	const [email, setEMail] = useState("")
-	const [invitationLink, setInvitationLink] = useState<string | null>(null)
-	const [inviteError, setInviteError] = useState<string | null>(null)
-	const [inviteToShare] = useInviteToShare({
-		onSuccess: (invitationLink) => {
-			if (invitationLink !== null) {
-				setInvitationLink(invitationLink)
-			}
-
-			message.success(`User ${email} succesfully invited`)
-
-			setEMail("")
-		},
-		onError: (err) => setInviteError(err.message),
-	})
-	const [revokeInvitation] = useRevokeInvitation({
-		onSuccess: () => {
-			setInvitationLink(null)
-
-			message.success(`User invitation successfully revoked`)
-		},
-	})
-	const [updatePermissions] = useUpdateShareMemberPermissions()
-
-	const onInviteClick = useCallback(() => {
-		inviteToShare({
-			input: {
-				shareID,
-				email,
-			},
-		})
-	}, [inviteToShare, shareID, email])
-
-	const onRevokeInvitationClick = useCallback(
-		(userID: string) => {
-			revokeInvitation({
-				input: {
-					shareID,
-					userID,
-				},
-			})
-		},
-		[revokeInvitation, shareID],
-	)
-
-	const onPermissionsValueChange = useCallback(
-		(user: ShareMember, permissions: string[]) => {
-			if (permissions.every((perm) => Permissions.ALL.includes(perm as Permission))) {
-				updatePermissions({
-					permissions,
-					userID: user.id,
-					shareID: user.shareID,
-				})
-			}
-		},
-		[updatePermissions],
-	)
-
-	if (error) return <div>Error</div>
-
-	return (
-		<>
-			<FormItemVertical label="Members">
-				<Table
-					dataSource={users || []}
-					pagination={false}
-					loading={isLoading}
-					scroll={{ y: 300 }}
-					rowKey={(user) => user.id}
-				>
-					<Column title="Name" dataIndex="name" key="name" />
-					<Column title="E-Mail" dataIndex="email" key="email" />
-					<Column title="Status" dataIndex="status" key="status" width={100} />
-					<Column
-						title="Actions"
-						key="actions"
-						width={120}
-						render={(_, user: ShareMember) => (
-							<>
-								{user.status === UserStatus.Pending && (
-									<Button type="link" onClick={() => onRevokeInvitationClick(user.id)}>
-										Revoke
-									</Button>
-								)}
-							</>
-						)}
-					/>
-					<Column
-						title="Permissions"
-						key="permissions"
-						render={(_, user: ShareMember) => {
-							return (
-								<EditableTagGroup
-									values={user.permissions}
-									placeholder="Permissions"
-									datasource={Permissions.ALL}
-									onValuesChange={(permissions) => onPermissionsValueChange(user, permissions)}
-								/>
-							)
-						}}
-					/>
-				</Table>
-			</FormItemVertical>
-			{invitationLink && (
-				<Alert
-					message={
-						<>
-							<span>Invitation link: </span>
-							<Text code>{invitationLink}</Text>
-						</>
-					}
-					type="success"
-					closable
-					onClose={() => setInvitationLink(null)}
-				/>
-			)}
-			{inviteError && (
-				<Alert
-					message={inviteError.replace("GraphQL error: ", "")}
-					type="error"
-					closable
-					onClose={() => setInviteError(null)}
-				/>
-			)}
-			<FormItemVertical label="E-Mail">
-				<Input
-					value={email}
-					type="email"
-					onChange={(e) => setEMail(e.target.value)}
-					placeholder="example@domain.com"
-					width={300}
-				/>
-				<Button type="dashed" onClick={onInviteClick}>
-					Invite
-				</Button>
-			</FormItemVertical>
-		</>
 	)
 }
